move submit button into shared a2 space
because we will eventually allow for multiple submission types
in an assignment we will need to be able to submit for multiple
type with a single press of the submit button.

for this purpose we are moving the submit button into a shared
space that is capable of managing the submission of all valid
types for an assignment.

this commit removes the upload and submit alert handling to be
added at a future time.

Test Plan:
* as a teacher, create an online_upload assignment
* as a student in a2, navigate to this assignment
* confirm you can:
** not see the submit button by default
** can add and remove attachments
** can see the pizza tracker update correctly with changes
** can see the submit button when attachments are added
** the button is removed correctly when attachments are
   removed
** can successfully submit to the assignment when the button
   is pressed

* as a teacher, create an online_text_entry assignment
* as a student in a2, navigate to this assignment
* confirm you can:
** not see the submit button by default
** can start an entry
** can edit an entry
** can delete an entry
** can see the pizza tracker update correctly with changes
** can see the submit button when a draft body is added
** the button is removed when the draft is cleared
Note: the submit button currently is not active for text

flag=assignments_2
fixes COMMS-2303

Change-Id: I02dd71ccd07646424bd2ad2bb3f522aceb58fd12
Reviewed-on: https://gerrit.instructure.com/205965
Reviewed-by: Landon Gilbert-Bland <lbland@instructure.com>
QA-Review: Landon Gilbert-Bland <lbland@instructure.com>
Tested-by: Jenkins
Product-Review: Ryan Norton <rnorton@instructure.com>