fix polyfill for old browsers
fixes: COREFE-272

core-js changed the name of all of their polyfills so they are not 
based on es6 or es7 but just es. 

This is to update our code to reflect that.

if you look at what's on prod
https://du11hjcvx0uqb.cloudfront.net/br/dist/ie11-polyfill-36fc8ad84a.js
vs what's on beta:
https://du11hjcvx0uqb.cloudfront.net/br/dist/ie11-polyfill-f3223dcd57.js
you can see that core-js stopped polyfilling some stuff.

so even thought this is specifically called the ie11-polyfill, and we
don't officially support ie11, we should still probably warmfix this so
that if there are any browsers that do need any of these polyfills they
still get it, out of an abundance of caution.

test plan:
* with this applied, view-source and then look for: "ie11-polyfill"
* copy and paste that url into a new tab. 
* it should have a bunch of stuff in it, like the example from prod 
  above and not be small like the one from beta above

Change-Id: Ic92ba4a31a739db7a50eebcfd7ac63d9ad8f6742
Reviewed-on: https://gerrit.instructure.com/208605
Tested-by: Jenkins
QA-Review: Ryan Shaw <ryan@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>