fix crash when text entry is submitted in a2
we are attempting to focus on a View component when we update
the text entry to display a submitted or graded submission.
this is resulting in the page crashing because it is a
non-focusable element.

this commit removes that focus while we decide how we want to
focus on higher-order components up the chain.

Test Plan:
* as a teacher in a2, create a text entry assignment
* as a student in a2, draft and submit to this assignment
* the submission should proceed and there should be no error
* submit a second attempt to this assignment
* click the previous button to return to the last submission
* the page should display without error

fixes COMMS-2474

Change-Id: I70da57c7f1ba121176fac14ec476e1346bba0311
Reviewed-on: https://gerrit.instructure.com/211340
Reviewed-by: Landon Gilbert-Bland <lbland@instructure.com>
Tested-by: Jenkins
QA-Review: Ryan Norton <rnorton@instructure.com>
Product-Review: Ryan Norton <rnorton@instructure.com>