perf: speed up discussions index js rendering
fixes: COREFE-346

This is an example of a pretty common problem with pages that use InstUI
components rendering slowly in canvas. I finally tracked down the “why”
behind it.

To reproduce, go to courses/x/discussion_topics in canvas in prod mode
if you use the devtools “Performance” panel to profile the page load,
you will see a ton of forced “Reactualate styles” and “layout”
eg: https://cl.ly/f6a3cbec5fb0

In this case, the reason why is that each of those rows has a component
that passes an explicit “theme” prop to InstUI’s themeable.
In the componentDidMount of Themeable, it calls applyTheme which calls
applyVariablesToNode which calls 
domNode.style.setProperty(propertyName, value), which means the DOM is
Now “dirty” and if anything asks to read a layout or style value, 
it must be recalculated.

And then, because some of those elements are also instUI Focusable 
components they also have a componentDidMount that calls findFocusable
which calls getComputedStyle(element) which forces style and layout to 
be recalculated on componentDidMount for every one of those components

So essentially because we are both reading and writing style related
stuff in every componentDidMount, it is causing a ton of 
“layout thrashing” and it has to recalculate for every one of them.

To fix it on this page, I just took out the 2 explicit “theme” props,
because it looks basically the same without them. But we should look
into instUI to see if there is something we can do with either Themeable
Or Focusable to make them not read/write synchronously so much to the 
DOM

Before: this took ~6 seconds on my laptop
After: this takes ~600ms

I made a screencast so you can see the problem more clearly:
https://instructure.slack.com/archives/CK5V51Y85/p1570214391290400

Test plan:
* go to /courses/x/discussion_topics in prod mode in canvas
* it should render a lot faster
* it should look exactly the same, except for a tiny bit more spacing
  On that “3 dot” menu as shown in the screencast

Change-Id: I197f236729f1ded2b68cf298c57ba542c5649999
Reviewed-on: https://gerrit.instructure.com/212161
Tested-by: Jenkins
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>