fix grade summary when streaming by wrapping in domready
Fixes: COREFE-373

Note: I’d love to add a spec for this but the only thing that could
have caught it was a selenium spec with streaming turned on.
Unfortunately though, James couldn’t figure out how to let streaming 
be turned on for our selenium tests so streaming is always off on
Jenkins when selenium runs. When we fix that, the existing selenium
tests for this page will catch things like this

Test plan:
* with streaming ON:
* go to the student grade summary page:
  (Eg: /courses/courseid/grades/studentid)
* you should not get a react error like:
  “Target container is not a DOM element.”
* the page should load fine

Change-Id: I467f6eaab7039d3e8fe5117747f76baa7f11dcdc
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/214065
Tested-by: Jenkins
Reviewed-by: James Williams <jamesw@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>