spec: remove New Quizzes Submission History spec
The feature flag, as it is currently written, cannot be turned off.
This spec is asserting on a state that can't be true, and is brittle
since it cares about the ordering that feature flags are checked in.

Test Plan
- View account feature flags as a site admin. Verify that the
  "New Quizzes Submission History" feature flag cannot be disabled.

Change-Id: Id2d27518df88ad790b7646b89e681d38f43b9435
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/214371
Reviewed-by: Spencer Olson <solson@instructure.com>
QA-Review: Gary Mei <gmei@instructure.com>
Product-Review: Gary Mei <gmei@instructure.com>
Tested-by: Jenkins