Distinguish course/account tool installs in ACTL
The AssignmentConfigurationToolLookup (ACTL) represents that a given
assignment is connected with a given plagiarism detection (LTI2) tool.
This is a soft link, not a direct "tool_proxy_id" link to the tool
installation, though, since we can uninstall and reinstall tools and we
want those connections (actually LESS subscriptions) to be
reestablished. We used to figure out that a ACTL belongs to a give tool
by matching up product and vendor codes; but if the user installs the
tool in both a Course context and an Account context, we can get
confused, and one tool from either context can delete the subscriptions
created with the tool in the other context.

The solution is to store the context_type (Course or Account) in the
ACTL when setting up an assignment with a plagiarism detection tool.

Test plan:
- instrument the locations in the canvas code where a LESS subscription
  is created and destroyed. You can comment out the calls and replace
  with "puts" statements in lib/lti/assignment_subscriptions_helper.rb:
    * replace `destroy_subscription` to not send to LESS, but log
      (append to a file is easiest to see) with the ID
    * replace `create_subscription` - make it log (append to a file)
      and return a fake id like `rand(100000).to_s
    * run both canvas web and jobs
- install a locally-running test tool (e.g.
  web.lti-originality-report-example.docker:3000/register)
  in a course context and add an assignment that uses the tool.
  (note: sometimes when adding a tool it didn't show up in the
  new assignment dropdown. restarting canvas may help)
- Check subscription id:
  `AssignmentConfigurationToolLookup.where(assignment_id: a_id`
  The ACTL should also have a context_type of "Course"
- install the test tool in a account context. check (by looking at ACTL)
  that the subscription ID did not change, and that no subscriptions were
  created or destroyed.
- delete the account-level test tool. observe that nothing changed.
  delete the course-level test tool, observe that the subscriptions
  have been deleted (looking at the logging you added).
- you can do other combinations like creating assignments with
  an account-level tool.
- Test "Blueprint course" syncing still works by creating a blueprint
  course, adding a tool and assignment with tool, making copies of the
  course, and syncing the copies. Check that there is still only one
  copy of the ACTL per course.

Fixes PLAT-5023

Change-Id: I6ded9ab91e223fc1cf6a9b7ed7f0df649823c95d
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/215171
Tested-by: Jenkins
Reviewed-by: Weston Dransfield <wdransfield@instructure.com>
QA-Review: Weston Dransfield <wdransfield@instructure.com>
Product-Review: Jesse Poulos <jpoulos@instructure.com>