spec: allow a tool to to hit LTI API in a provider state
This stubs a few parts of our JWT verification process so that
an expired JWT can be used to make API calls when the contract
test build runs. The JWT still needs to be valid, aside from
the iat and exp timestamps, and should be signed with the key
in the live-events-lti repository.

test plan:
- Set up the live-events-subscription-service locally
- Run the contract tests for the live-events-lti repo
  and save the generated JSON file somewhere
- In canvas-lms, run
  rails pact:verify:at[./path_to_contracts.json]
- It should pass

fixes PLAT-5101

Change-Id: Iff5a6a91aa2ad9868511d3396c73d8225587f640
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/216596
Tested-by: Jenkins
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Marc Phillips <mphillips@instructure.com>
QA-Review: Tucker Mcknight <tmcknight@instructure.com>
Product-Review: Tucker Mcknight <tmcknight@instructure.com>