pass x-requested-with header for `fetch` requests
rails uses this header as a heuristic to know if a request is an xhr
request. Specifically here's the method:

    def xml_http_request?
      get_header("HTTP_X_REQUESTED_WITH") =~ /XMLHttpRequest/i
    end
    alias :xhr? :xml_http_request?

unfortunately this is not really a standard, just something jquery did
that rails uses. so we need to emulate it whenever we make requests via
other means. indeed, we already do this for axios.

we may want to consider making a `fetch` wrapper so future uses remember
to do this, but updating our default fetch options is a good start.

fixes CNVS-47648

test plan:
- make sure pageviews are enabled:
    `Setting.set('enable_page_views', 'true')`
- run rails, visit a page, and wait until the page checks unread_status
- make sure a page view is not logged for that request

Change-Id: I2b4b928b6d2c5a51e49045329026eeb23f1d743b
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/217855
Reviewed-by: Ryan Shaw <ryan@instructure.com>
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Tested-by: Jenkins
QA-Review: Steve Shepherd <sshepherd@instructure.com>
Product-Review: Simon Williams <simon@instructure.com>