Requery rce tray content when filter changes
When the CanvasContentTray's filter changed the contextType,
the query logic reuses the existing query result if one exists.
If the user switches context, changes sort, the switches context back,
the previous result in the incorrect sort order is displayed.

To reuse the existing query, we would have to track the sort params
per contextType, which would require it be in the redux store, which
would mean a lot of changes to actions, reducers and UI components
to track that. The effort (at this time) is not worth the reward and
we'll just requery the data if either the contextType or the sort
changes. The data should be in the browser's cache anyway.

The other change here is to delegate the logic for reexecuting the
initial query for the first page of results to the useIncrementalLoading
custom hook, removing it from each of the CanvasContentTray's panel
components.

closes LA-419
flag=rce_enhancements

test plan:
  - in a course that has some files, images, and media
  - for a user that has some files, images, and media
  - create a new Assignment or Announcement (to get an RCE)
  - from the toolbar, select Documents > Course Documents
  > expect the tray to open, displaying the course docs
    in descending order by date
  > expect there to be only 2 network queries for documents
    (one the method is OPTIONS, one GET)
  - switch sort to "Alphabetical"
  > expect the file to be displayed in alphabetical order
  - switch from Course Files to User Files
  > expect an alphabetical list of the user's files
  - switch sort to "Date Added"
  > expect the list to be displayed by date, descending
  - switch back to "Course Files"
  > expect the list of course files to also be displayed
    by date, descending

  - repeat for images and media

  - select Links > Course Links from the toolbar
  - expand any of the sub-sections (e.g. Pages, Assignments)
  > expect the  section to be populated

Change-Id: I592d5cec32f98d5a243386760a5a37f3e207d21c
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/222221
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jon Willesen <jonw+gerrit@instructure.com>
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>