Fix scopes on creation of lti key
This was introduced in
https://gerrit.instructure.com/c/canvas-lms/+/219146
where the code to do customizations was removed. Turns
out that the customization page saved the scopes on the
developer key. This makes it save the scopes from the
tool_configuration settings.

closes PLAT-5285

Test Plan:
 - attempt to make a dev key in any manner
 - check in the console that the newly created dev key
   has scopes attached

Change-Id: I76cd42d15daf6f69bc04ad90768055254b5850b3
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/222646
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Tucker Mcknight <tmcknight@instructure.com>
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
Product-Review: Marc Phillips <mphillips@instructure.com>