create button for course level notifications
Adds a button to the course homepage sidebar to allow a user to
navigate to the Course Notification Settings page. Requires the
mute_notifications_by_course feature flag to be enabled, and for
the user to have an enrollment in the course.

TEST PLAN:
With the feature flag enabled
AND a user with an enrollment in a course
 - The button should be visible on the course homepage
 - The button should navigate to courses/<id>?view=notifications
With the feature flag enabled
AND a user NOT enrolled in the course (such as an admin)
 - The button should NOT be visible
With the feature flag disabled
AND any user
 - The button should NOT be visible

closes KNO-226
flag=mute_notifications_by_course

Change-Id: Ic4b3791050085ec06a5003516d8193854e6a528c
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/226091
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Matthew Lemon <mlemon@instructure.com>
Reviewed-by: Johnny Le <jle@instructure.com>
QA-Review: Johnny Le <jle@instructure.com>
Product-Review: Johnny Le <jle@instructure.com>