spec: expect user to be included
expecting user to be the first in enrollments leads
to flakey results. Instead check for the user to be
included in the enrollments, no requirement for this
user to be the first.

flag = none
Test Plan:
passes jenkins specifically FSC build

Change-Id: Ia067daf8bb8d8ad2dec751743b1e87505887ef6e
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/230918
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Robin Kuss <rkuss@instructure.com>
QA-Review: James Butters <jbutters@instructure.com>
Product-Review: James Butters <jbutters@instructure.com>